+++ bundles/framework/mapmodule-plugin/event/AfterAddMarkerEvent.js
@@ -25,6 +25,15 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.event.AfterAddMarkerEv
         getID: function () {
             return this._id;
         }
+        /**
+         * Serialization for RPC
+         * @return {Object} object has key id which has the marker id of the clicked
+         */
+        getParams: function () {
+            return {
+                id: this.getID()
+            };
+        }
     }, {
         /**
          * @property {String[]} protocol array of superclasses as {String}